@@ -7,6 +7,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
@@ -28,7 +29,7 @@ public class SecurityConfig extends VaadinWebSecurity {
   @Override
   protected void configure(HttpSecurity http) throws Exception {
     // Authorize access to /images/ without authentication
-    http.authorizeRequests().antMatchers("/images/**").permitAll();
+    http.authorizeHttpRequests().requestMatchers("/images/**").permitAll();
     // Set default security policy that permits Vaadin internal requests and
     // denies all other
     super.configure(http);
